requirements.txt
numpy==1.19.4
Flask_Cors==3.0.10
Werkzeug==1.0.1
Flask_WTF==0.15.1
Flask==1.1.2
Flask_Login==0.5.0
Flask_RESTful==0.3.9
SQLAlchemy==1.4.9
WTForms==2.3.3
PyMySQL==0.9.3
Flask_SQLAlchemy==2.5.1
PyJWT==2.1.0
python-dotenv==0.18.0
email-validator==1.1.3
idna==3.2
dnspython==2.1.0
